pyproject.toml
[tool.poetry]
name = "psql-backup"
version = "0.1.0"
description = "an app to take backups of my psql databases on regular intervals"
authors = ["deadaf <gahlyanrohit18@gmail.com>"]
readme = "README.md"
packages = [{include = "psql_backup"}]

[tool.poetry.dependencies]
python = "^3.10"
python-decouple = "^3.7"
"mega.py" = "^1.0.8"
asyncpg = "^0.27.0"



[tool.black]
line-length = 122

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
